# docs/swagger.yaml
basePath: /v1
definitions:
...
  models.Scheduler:
    properties:
      amount:
        type: integer
      base:
        type: string
      capacity:
        type: integer
      classId:
        type: string
      className:
        type: string
      customerId:
        type: string
      doorId:
        description: DoorID          *string `gorm:"type:varchar(256);" json:"doorId"`
        type: integer
      employeeId:
        type: string
      endClassTime:
        type: integer
      endDate:
        type: string
      id:
        type: integer
      lecturerId:
        type: string
      lecturerName:
        type: string
      role:
        description: value in ["employee", "student", "customer"]
        type: string
      roomId:
        type: string
      roomName:
        type: string
      roomRow:
        type: string
      startClassTime:
        type: integer
      startDate:
        type: string
      status:
        type: string
      studentId:
        type: string
      userId:
        type: string
      weekDay:
        type: integer
    type: object
...
  models.SwagCreateScheduler:
    properties:
      amount:
        type: integer
      base:
        type: string
      capacity:
        type: integer
      classId:
        type: string
      className:
        type: string
      endClassTime:
        type: integer
      endDate:
        type: string
      lecturerId:
        type: string
      lecturerName:
        type: string
      roomId:
        type: string
      roomName:
        type: string
      roomRow:
        type: string
      startClassTime:
        type: integer
      startDate:
        type: string
      status:
        type: string
      weekDay:
        type: integer
    type: object
...
  models.SwagUpdateScheduler:
    properties:
      amount:
        type: integer
      base:
        type: string
      capacity:
        type: integer
      classId:
        type: string
      className:
        type: string
      endClassTime:
        type: integer
      endDate:
        type: string
      id:
        type: integer
      lecturerId:
        type: string
      lecturerName:
        type: string
      roomId:
        type: string
      roomName:
        type: string
      roomRow:
        type: string
      startClassTime:
        type: integer
      startDate:
        type: string
      status:
        type: string
      weekDay:
        type: integer
    type: object
...
  models.UpdateScheduler:
    properties:
      amount:
        type: integer
      base:
        type: string
      capacity:
        type: integer
      classId:
        type: string
      className:
        type: string
      customerId:
        type: string
      doorId:
        description: DoorID          *string `gorm:"type:varchar(256);" json:"doorId"`
        type: integer
      doorlockAddress:
        type: string
      employeeId:
        type: string
      endClassTime:
        type: integer
      endDate:
        type: string
      id:
        type: integer
      lecturerId:
        type: string
      lecturerName:
        type: string
      role:
        description: value in ["employee", "student", "customer"]
        type: string
      roomId:
        type: string
      roomName:
        type: string
      roomRow:
        type: string
      startClassTime:
        type: integer
      startDate:
        type: string
      status:
        type: string
      studentId:
        type: string
      userId:
        type: string
      weekDay:
        type: integer
    required:
    - doorlockAddress
    - userId
    type: object
...
  utils.ErrorResponse:
    properties:
      errorMsg:
        type: string
      msg:
        type: string
      statusCode:
        type: integer
    type: object
host: http://iot.hcmue.space:8002
info:
  contact:
    name: Mr. Khai
  description: This is API document for DMS backend server
  title: Backend API
  version: "1.0"
paths:
  /v1//scheduler/excel:
    patch:
      consumes:
      - application/json
      description: Update Scheduler base on Excel
      parameters:
      - description: Fields need to append a scheduler
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.SwagUpdateScheduler'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Scheduler'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Update Scheduler Base On Excel
    post:
      consumes:
      - application/json
      description: Append Scheduler base on Excel
      parameters:
      - description: Fields need to append a scheduler
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.SwagCreateScheduler'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Scheduler'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Append Scheduler Base On Excel
...
  /v1/scheduler:
    delete:
      consumes:
      - application/json
      description: Delete scheduler using "id" field. Send deleted info to MQTT broker
      parameters:
      - description: Scheduler ID
        in: body
        name: data
        required: true
        schema:
          allOf:
          - type: object
          - properties:
              id:
                type: integer
            type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Delete Scheduler By ID
    patch:
      consumes:
      - application/json
      description: Update scheduler, must have "id" field. Send updated info to MQTT
        broker
      parameters:
      - description: Fields need to update a scheduler
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.UpdateScheduler'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Update Scheduler By ID
    post:
      consumes:
      - application/json
      description: Create scheduler
      parameters:
      - description: Fields need to create a scheduler
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/models.SwagCreateScheduler'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Scheduler'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Create Scheduler
  /v1/scheduler/{id}:
    get:
      description: find scheduler info by scheduler id
      parameters:
      - description: Scheduler ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Scheduler'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Find Scheduler By ID
  /v1/schedulers:
    get:
      description: find all scheduler info
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              items:
                $ref: '#/definitions/models.Scheduler'
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.ErrorResponse'
      summary: Find All Scheduler
...
swagger: "2.0"